// frontend/src/services/generated/model/getRandomQuestionsByDifficultyDifficulty.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type GetRandomQuestionsByDifficultyDifficulty = typeof GetRandomQuestionsByDifficultyDifficulty[keyof typeof GetRandomQuestionsByDifficultyDifficulty];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetRandomQuestionsByDifficultyDifficulty = {
  EASY: 'EASY',
  MEDIUM: 'MEDIUM',
  HARD: 'HARD',
} as const;