// frontend/src/services/generated/model/getRandomQuestionsByDifficultyParams.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { GetRandomQuestionsByDifficultyDifficulty } from './getRandomQuestionsByDifficultyDifficulty';

export type GetRandomQuestionsByDifficultyParams = {
/**
 * Úroveň obtížnosti otázek. Validní hodnoty: EASY, MEDIUM, HARD. Pokud obtížnost neexistuje, vrátí se prázdný seznam.
 */
difficulty: GetRandomQuestionsByDifficultyDifficulty;
/**
 * Maximální počet otázek k vrácení. Musí být kladné číslo a max 100.
 */
limit: number;
};